[TASK] rename existing boilerplate directories
I decided to add a numerical postfix in order to make the decision on how
to name new boilerplates more easy. However, one should add a useful and
short DESCRIPTION file when adding new boilerplates because a basic
category plus number will not tell much about the structure and features
of a boilerplate.