[BUGFIX] fix search & replace on Mac OS / BSD
The used sed command was working with GNU sed only. An additional check was
added to make sure BSD sed syntax is used on Mac OS and FreeBSD. Thanks to
martin2110 for reporting this.

Resolves: Issue #3

Additional change: check order to detect an invalid boilerplate nr. choice.
This should prevent a bash errors when a user insert values like "0_foobar"
instead of a numeric value.