[TASK] Fix "-o" behavior ("no GitHub Org/Bitbucket Team"); Remove Bashisms
There was no possibility to submit that there is *no* GitHub Organisation or
Bitbucket Team to use. This is now possible by using "-o false" or "-o none".
See the help for details (call the script with -h).

The whole script was re-rewritten to make it compatible to more shells (Bash,
Dash). Therefore, typical Bashisms were removed. The quoting style was also
changed: All strings that do not contain variables should be enclosed in single
quotes now. Double quotes are used when variable interpolation is required.
"echo" was replaced with "printf" (it is more portable, there are just to many
"traditional" versions of the echo utility out there).